Added OIDC based authorization feature

+++ lib/Controller/OpenProjectAPIController.php
@@ -86,6 +86,29 @@ class OpenProjectAPIController extends Controller {
 		$this->logger = $logger;
 	}
 
+	private function validatePreRequestConditions(): array {
+		$authMethod = $this->config->getAppValue(Application::APP_ID, 'authorization_method', '');
+		if ($authMethod === OpenProjectAPIService::AUTH_METHOD_OAUTH && $this->accessToken === '') {
+			return [
+				'status' => false,
+				'result' => new DataResponse('', Http::STATUS_UNAUTHORIZED)
+			];
+		} elseif ($authMethod === OpenProjectAPIService::AUTH_METHOD_OIDC &&
+			$this->openprojectAPIService->getOIDCBasedTokenForTheTargetedAudienceClient() === null
+		) {
+			return [
+				'status' => false,
+				'result' => new DataResponse('', Http::STATUS_UNAUTHORIZED)
+			];
+		} elseif (!OpenProjectAPIService::validateURL($this->openprojectUrl)) {
+			return [
+				'status' => false,
+				'result' => new DataResponse('', Http::STATUS_BAD_REQUEST)
+			];
+		}
+		return ['status' => true, 'result' => null];
+	}
+
 	/**
 	 * get openproject instance URL
 	 * @NoAdminRequired
@@ -95,7 +118,6 @@ class OpenProjectAPIController extends Controller {
 	public function getOpenProjectUrl(): DataResponse {
 		return new DataResponse($this->openprojectUrl);
 	}
-
 	/**
 	 * get openproject user avatar
 	 * @NoAdminRequired
@@ -126,10 +148,9 @@ class OpenProjectAPIController extends Controller {
 	 * @return DataResponse
 	 */
 	public function getNotifications(): DataResponse {
-		if ($this->accessToken === '') {
-			return new DataResponse('', Http::STATUS_UNAUTHORIZED);
-		} elseif (!OpenProjectAPIService::validateURL($this->openprojectUrl)) {
-			return new DataResponse('', Http::STATUS_BAD_REQUEST);
+		$validatePreRequestResult = $this->validatePreRequestConditions();
+		if (!$validatePreRequestResult['status']) {
+			return $validatePreRequestResult['result'];
 		}
 		$result = $this->openprojectAPIService->getNotifications($this->userId);
 		if (!isset($result['error'])) {
@@ -160,10 +181,9 @@ class OpenProjectAPIController extends Controller {
 		?int $fileId = null,
 		bool $isSmartPicker = false
 	): DataResponse {
-		if ($this->accessToken === '') {
-			return new DataResponse('', Http::STATUS_UNAUTHORIZED);
-		} elseif (!OpenProjectAPIService::validateURL($this->openprojectUrl)) {
-			return new DataResponse('', Http::STATUS_BAD_REQUEST);
+		$validatePreRequestResult = $this->validatePreRequestConditions();
+		if (!$validatePreRequestResult['status']) {
+			return $validatePreRequestResult['result'];
 		}
 		// when the search is done through smart picker we don't want to check if the work package is linkable
 		$result = $this->openprojectAPIService->searchWorkPackage(
@@ -198,12 +218,10 @@ class OpenProjectAPIController extends Controller {
 	 * @return DataResponse
 	 */
 	public function linkWorkPackageToFile(array $values): DataResponse {
-		if ($this->accessToken === '') {
-			return new DataResponse('', Http::STATUS_UNAUTHORIZED);
-		} elseif (!OpenProjectAPIService::validateURL($this->openprojectUrl)) {
-			return new DataResponse('', Http::STATUS_BAD_REQUEST);
+		$validatePreRequestResult = $this->validatePreRequestConditions();
+		if (!$validatePreRequestResult['status']) {
+			return $validatePreRequestResult['result'];
 		}
-
 		try {
 			$result = $this->openprojectAPIService->linkWorkPackageToFile(
 				$values,
@@ -227,10 +245,9 @@ class OpenProjectAPIController extends Controller {
 	 * @return DataResponse
 	 */
 	public function markNotificationAsRead(int $workpackageId) {
-		if ($this->accessToken === '') {
-			return new DataResponse('', Http::STATUS_UNAUTHORIZED);
-		} elseif (!OpenProjectAPIService::validateURL($this->openprojectUrl)) {
-			return new DataResponse('', Http::STATUS_BAD_REQUEST);
+		$validatePreRequestResult = $this->validatePreRequestConditions();
+		if (!$validatePreRequestResult['status']) {
+			return $validatePreRequestResult['result'];
 		}
 		try {
 			$result = $this->openprojectAPIService->markAllNotificationsOfWorkPackageAsRead(
@@ -257,12 +274,10 @@ class OpenProjectAPIController extends Controller {
 	 * @return DataResponse
 	 */
 	public function getWorkPackageFileLinks(int $id): DataResponse {
-		if ($this->accessToken === '') {
-			return new DataResponse('', Http::STATUS_UNAUTHORIZED);
-		} elseif (!OpenProjectAPIService::validateURL($this->openprojectUrl)) {
-			return new DataResponse('', Http::STATUS_BAD_REQUEST);
+		$validatePreRequestResult = $this->validatePreRequestConditions();
+		if (!$validatePreRequestResult['status']) {
+			return $validatePreRequestResult['result'];
 		}
-
 		try {
 			$result = $this->openprojectAPIService->getWorkPackageFileLinks(
 				$id,
@@ -284,12 +299,10 @@ class OpenProjectAPIController extends Controller {
 	 * @return DataResponse
 	 */
 	public function deleteFileLink(int $id): DataResponse {
-		if ($this->accessToken === '') {
-			return new DataResponse('', Http::STATUS_UNAUTHORIZED);
-		} elseif (!OpenProjectAPIService::validateURL($this->openprojectUrl)) {
-			return new DataResponse('', Http::STATUS_BAD_REQUEST);
+		$validatePreRequestResult = $this->validatePreRequestConditions();
+		if (!$validatePreRequestResult['status']) {
+			return $validatePreRequestResult['result'];
 		}
-
 		try {
 			$result = $this->openprojectAPIService->deleteFileLink(
 				$id,
@@ -315,12 +328,10 @@ class OpenProjectAPIController extends Controller {
 	 * @return DataResponse
 	 */
 	public function getOpenProjectWorkPackageStatus(string $id): DataResponse {
-		if ($this->accessToken === '') {
-			return new DataResponse('', Http::STATUS_UNAUTHORIZED);
-		} elseif (!OpenProjectAPIService::validateURL($this->openprojectUrl)) {
-			return new DataResponse('', Http::STATUS_BAD_REQUEST);
+		$validatePreRequestResult = $this->validatePreRequestConditions();
+		if (!$validatePreRequestResult['status']) {
+			return $validatePreRequestResult['result'];
 		}
-
 		$result = $this->openprojectAPIService->getOpenProjectWorkPackageStatus(
 			$this->userId, $id
 		);
@@ -342,12 +353,10 @@ class OpenProjectAPIController extends Controller {
 	 * @return DataResponse
 	 */
 	public function getOpenProjectWorkPackageType(string $id): DataResponse {
-		if ($this->accessToken === '') {
-			return new DataResponse('', Http::STATUS_UNAUTHORIZED);
-		} elseif (!OpenProjectAPIService::validateURL($this->openprojectUrl)) {
-			return new DataResponse('', Http::STATUS_BAD_REQUEST);
+		$validatePreRequestResult = $this->validatePreRequestConditions();
+		if (!$validatePreRequestResult['status']) {
+			return $validatePreRequestResult['result'];
 		}
-
 		$result = $this->openprojectAPIService->getOpenProjectWorkPackageType(
 			$this->userId, $id
 		);
@@ -365,10 +374,9 @@ class OpenProjectAPIController extends Controller {
 	 * @return DataResponse
 	 */
 	public function getAvailableOpenProjectProjects(?string $searchQuery = null): DataResponse {
-		if ($this->accessToken === '') {
-			return new DataResponse('', Http::STATUS_UNAUTHORIZED);
-		} elseif (!OpenProjectAPIService::validateURL($this->openprojectUrl)) {
-			return new DataResponse('', Http::STATUS_BAD_REQUEST);
+		$validatePreRequestResult = $this->validatePreRequestConditions();
+		if (!$validatePreRequestResult['status']) {
+			return $validatePreRequestResult['result'];
 		}
 		try {
 			$result = $this->openprojectAPIService->getAvailableOpenProjectProjects($this->userId, $searchQuery);
@@ -417,10 +425,9 @@ class OpenProjectAPIController extends Controller {
 	 * @return DataResponse
 	 */
 	public function getOpenProjectWorkPackageForm(string $projectId, array $body): DataResponse {
-		if ($this->accessToken === '') {
-			return new DataResponse('', Http::STATUS_UNAUTHORIZED);
-		} elseif (!OpenProjectAPIService::validateURL($this->openprojectUrl)) {
-			return new DataResponse('', Http::STATUS_BAD_REQUEST);
+		$validatePreRequestResult = $this->validatePreRequestConditions();
+		if (!$validatePreRequestResult['status']) {
+			return $validatePreRequestResult['result'];
 		}
 		try {
 			$result = $this->openprojectAPIService->getOpenProjectWorkPackageForm($this->userId, $projectId, $body);
@@ -439,10 +446,9 @@ class OpenProjectAPIController extends Controller {
 	 * @return DataResponse
 	 */
 	public function getAvailableAssigneesOfAProject(string $projectId): DataResponse {
-		if ($this->accessToken === '') {
-			return new DataResponse('', Http::STATUS_UNAUTHORIZED);
-		} elseif (!OpenProjectAPIService::validateURL($this->openprojectUrl)) {
-			return new DataResponse('', Http::STATUS_BAD_REQUEST);
+		$validatePreRequestResult = $this->validatePreRequestConditions();
+		if (!$validatePreRequestResult['status']) {
+			return $validatePreRequestResult['result'];
 		}
 		try {
 			$result = $this->openprojectAPIService->getAvailableAssigneesOfAProject($this->userId, $projectId);
@@ -488,10 +494,9 @@ class OpenProjectAPIController extends Controller {
 	 * @return DataResponse
 	 */
 	public function createWorkPackage(array $body): DataResponse {
-		if ($this->accessToken === '') {
-			return new DataResponse('', Http::STATUS_UNAUTHORIZED);
-		} elseif (!OpenProjectAPIService::validateURL($this->openprojectUrl)) {
-			return new DataResponse('', Http::STATUS_BAD_REQUEST);
+		$validatePreRequestResult = $this->validatePreRequestConditions();
+		if (!$validatePreRequestResult['status']) {
+			return $validatePreRequestResult['result'];
 		}
 		// we don't want to check if all the data in the body is set or not because
 		// that calculation will be done by the openproject api itself
@@ -517,10 +522,9 @@ class OpenProjectAPIController extends Controller {
 	 * @return DataResponse
 	 */
 	public function getOpenProjectConfiguration(): DataResponse {
-		if ($this->accessToken === '') {
-			return new DataResponse('', Http::STATUS_UNAUTHORIZED);
-		} elseif (!OpenProjectAPIService::validateURL($this->openprojectUrl)) {
-			return new DataResponse('', Http::STATUS_BAD_REQUEST);
+		$validatePreRequestResult = $this->validatePreRequestConditions();
+		if (!$validatePreRequestResult['status']) {
+			return $validatePreRequestResult['result'];
 		}
 		try {
 			$result = $this->openprojectAPIService->getOpenProjectConfiguration($this->userId);
